% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChebiEntry.R
\name{ChebiEntry}
\alias{ChebiEntry}
\title{ChEBI entry class.}
\description{
This is the entry class for ChEBI database.
}
\examples{
\donttest{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Load package definitions:
if ( ! mybiodb$getDbsInfo()$isDefined('chebi')) {
    mybiodb$loadDefinitions(system.file("definitions.yml", package='biodbChebi'))
}

# Check if ChEBI SOAP service is available
if (tryCatch({
    url <- "https://www.ebi.ac.uk/webservices/chebi/2.0/webservice?wsdl"
    res <- RCurl::getURL(url, .opts = list(timeout = 5, connecttimeout = 5))
    grepl("^\\\\s*<\\\\?xml|<wsdl:definitions", res)
}, error = function(e) FALSE)) {
    # Create a connector to ChEBI
    conn <- mybiodb$getFactory()$createConn('chebi')

    # Get an entry
    e <- conn$getEntry('15440')
}

# Terminate instance.
mybiodb$terminate()
}

}
\section{Super classes}{
\code{\link[biodb:BiodbEntry]{biodb::BiodbEntry}} -> \code{\link[biodb:BiodbXmlEntry]{biodb::BiodbXmlEntry}} -> \code{ChebiEntry}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ChebiEntry-clone}{\code{ChebiEntry$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="appendFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-appendFieldValue'><code>biodb::BiodbEntry$appendFieldValue()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="cloneInstance"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-cloneInstance'><code>biodb::BiodbEntry$cloneInstance()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="computeFields"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-computeFields'><code>biodb::BiodbEntry$computeFields()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="fieldHasBasicClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-fieldHasBasicClass'><code>biodb::BiodbEntry$fieldHasBasicClass()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getBiodb"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getBiodb'><code>biodb::BiodbEntry$getBiodb()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getDbClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getDbClass'><code>biodb::BiodbEntry$getDbClass()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getField'><code>biodb::BiodbEntry$getField()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldCardinality"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldCardinality'><code>biodb::BiodbEntry$getFieldCardinality()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldClass"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldClass'><code>biodb::BiodbEntry$getFieldClass()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldDef"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldDef'><code>biodb::BiodbEntry$getFieldDef()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldNames"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldNames'><code>biodb::BiodbEntry$getFieldNames()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldValue'><code>biodb::BiodbEntry$getFieldValue()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsAsDataframe"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsAsDataframe'><code>biodb::BiodbEntry$getFieldsAsDataframe()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsAsJson"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsAsJson'><code>biodb::BiodbEntry$getFieldsAsJson()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getFieldsByType"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getFieldsByType'><code>biodb::BiodbEntry$getFieldsByType()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getId"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getId'><code>biodb::BiodbEntry$getId()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getName"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getName'><code>biodb::BiodbEntry$getName()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="getParent"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-getParent'><code>biodb::BiodbEntry$getParent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="hasField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-hasField'><code>biodb::BiodbEntry$hasField()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="isNew"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-isNew'><code>biodb::BiodbEntry$isNew()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="makesRefToEntry"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-makesRefToEntry'><code>biodb::BiodbEntry$makesRefToEntry()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="parentIsAConnector"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-parentIsAConnector'><code>biodb::BiodbEntry$parentIsAConnector()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="parseContent"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-parseContent'><code>biodb::BiodbEntry$parseContent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="print"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-print'><code>biodb::BiodbEntry$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="removeField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-removeField'><code>biodb::BiodbEntry$removeField()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="setField"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-setField'><code>biodb::BiodbEntry$setField()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbEntry" data-id="setFieldValue"><a href='../../biodb/html/BiodbEntry.html#method-BiodbEntry-setFieldValue'><code>biodb::BiodbEntry$setFieldValue()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="biodb" data-topic="BiodbXmlEntry" data-id="initialize"><a href='../../biodb/html/BiodbXmlEntry.html#method-BiodbXmlEntry-initialize'><code>biodb::BiodbXmlEntry$initialize()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ChebiEntry-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ChebiEntry-clone}{}}}
\subsection{\code{ChebiEntry$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ChebiEntry$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
